Add files via upload
An international online store launched an A/B test, but analysts left it unfinished. I need to see whether the A/B test was carried out correctly and analyze it's results. I have only the technical specifications and the test results. The A/B test is dedicated to testing changes related to the introduction of an improved recommendation system
To do so:

I need to see whether within 14 days of signing up, users will show better conversion into product page views (the product_page event), product card views (product_card) and purchases (purchase).